fix: remove journey details page
