+++ apps/journeys-admin/pages/journeys/[journeyId].tsx
@@ -10,16 +10,19 @@ import { NextSeo } from 'next-seo'
 import { ReactElement } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
+import { ActiveJourneyEditContent } from '@core/journeys/ui/EditorProvider'
 import { JOURNEY_FIELDS } from '@core/journeys/ui/JourneyProvider/journeyFields'
 
 import { ACCEPT_ALL_INVITES } from '..'
 import { AcceptAllInvites } from '../../__generated__/AcceptAllInvites'
-import { GetJourney } from '../../__generated__/GetJourney'
+import {
+  GetJourney,
+  GetJourney_journey as Journey
+} from '../../__generated__/GetJourney'
 import { UserJourneyOpen } from '../../__generated__/UserJourneyOpen'
-import { JourneyInvite } from '../../src/components/JourneyInvite/JourneyInvite'
-import { JourneyView } from '../../src/components/JourneyView'
-import { Menu } from '../../src/components/JourneyView/Menu'
+import { Editor } from '../../src/components/Editor'
+import { EditToolbar } from '../../src/components/Editor/EditToolbar'
+import { JourneyEdit } from '../../src/components/Editor/JourneyEdit'
 import { PageWrapper } from '../../src/components/PageWrapper'
 import { initAndAuthApp } from '../../src/libs/initAndAuthApp'
 import { useInvalidJourneyRedirect } from '../../src/libs/useInvalidJourneyRedirect/useInvalidJourneyRedirect'
@@ -41,49 +44,40 @@ export const USER_JOURNEY_OPEN = gql`
   }
 `
 
-function JourneyIdPage(): ReactElement {
+function JourneyEditPage(): ReactElement {
   const { t } = useTranslation('apps-journeys-admin')
   const router = useRouter()
   const AuthUser = useAuthUser()
-  const { data, error } = useQuery<GetJourney>(GET_JOURNEY, {
+  const { data } = useQuery<GetJourney>(GET_JOURNEY, {
     variables: { id: router.query.journeyId }
   })
-
   useInvalidJourneyRedirect(data)
 
   return (
     <>
-      {error == null && (
-        <>
-          <NextSeo
-            title={data?.journey?.title ?? t('Journey')}
-            description={data?.journey?.description ?? undefined}
-          />
-          <JourneyProvider
-            value={{
-              journey: data?.journey ?? undefined,
-              variant: 'admin'
-            }}
-          >
-            <PageWrapper
-              title={t('Journey Details')}
-              showDrawer
-              backHref="/"
-              menu={<Menu />}
-              authUser={AuthUser}
-            >
-              <JourneyView journeyType="Journey" />
-            </PageWrapper>
-          </JourneyProvider>
-        </>
-      )}
-      {error?.graphQLErrors[0].message ===
-        'user is not allowed to view journey' && (
-        <>
-          <NextSeo title={t('Access Denied')} />
-          <JourneyInvite journeyId={router.query.journeyId as string} />
-        </>
-      )}
+      <NextSeo
+        title={
+          data?.journey?.title != null
+            ? t('Edit {{title}}', { title: data.journey.title })
+            : t('Edit Journey')
+        }
+        description={data?.journey?.description ?? undefined}
+      />
+      <Editor
+        journey={data?.journey ?? undefined}
+        selectedStepId={router.query.stepId as string | undefined}
+        view={router.query.view as ActiveJourneyEditContent | undefined}
+      >
+        <PageWrapper
+          title={data?.journey?.title ?? t('Edit Journey')}
+          showDrawer
+          backHref="/"
+          menu={<EditToolbar />}
+          authUser={AuthUser}
+        >
+          <JourneyEdit />
+        </PageWrapper>
+      </Editor>
     </>
   )
 }
@@ -101,15 +95,42 @@ export const getServerSideProps = withAuthUserTokenSSR({
 
   if (redirect != null) return { redirect }
 
-  await Promise.all([
-    apolloClient.mutate<AcceptAllInvites>({
-      mutation: ACCEPT_ALL_INVITES
-    }),
-    apolloClient.mutate<UserJourneyOpen>({
-      mutation: USER_JOURNEY_OPEN,
-      variables: { id: query?.journeyId }
+  await apolloClient.mutate<AcceptAllInvites>({
+    mutation: ACCEPT_ALL_INVITES
+  })
+
+  let journey: Journey | null
+  try {
+    const { data } = await apolloClient.query<GetJourney>({
+      query: GET_JOURNEY,
+      variables: {
+        id: query?.journeyId
+      }
     })
-  ])
+
+    journey = data?.journey
+  } catch (error) {
+    return {
+      redirect: {
+        permanent: false,
+        destination: `/journeys/${query?.journeyId as string}`
+      }
+    }
+  }
+
+  if (journey?.template === true) {
+    return {
+      redirect: {
+        permanent: false,
+        destination: `/publisher/${journey?.id}/edit`
+      }
+    }
+  }
+
+  await apolloClient.mutate<UserJourneyOpen>({
+    mutation: USER_JOURNEY_OPEN,
+    variables: { id: query?.journeyId }
+  })
 
   return {
     props: {
@@ -121,4 +142,4 @@ export const getServerSideProps = withAuthUserTokenSSR({
 
 export default withAuthUser({
   whenUnauthedAfterInit: AuthAction.REDIRECT_TO_LOGIN
-})(JourneyIdPage)
+})(JourneyEditPage)
